remove DBUS permissions for the KDE tray
Those seem overly abusive. I haven't heard reports of the tray working
or not, so let's be on the careful side for now. If that fails, we can
revert this.

I used this to do an inventory:

https://github.com/search?q=org%3Aflathub+--own-name%3Dorg.kde&type=code

Some also use `--talk-name` but less frequently, and we already do
have such an entry.

Most people seem to do the wildcard, but according to this:

https://github.com/flathub/flatpak-builder-lint/pull/114

> As of Qt 6.2 the tray works as expected without extra
> permissions. This version should be in all supported KDE runtimes.
>
> Using org.kde.* is not a correct permission and is a security risk.

So let's just drop this completely and see what happens.